add argparse as zig module

+++ src/model/ecs/components/basaltic_components_planes.c
@@ -372,7 +372,7 @@ const char *plane_getCellLifezoneName(const Plane *plane, const Climate *climate
     s32 biotemp = plane_GetCellBiotemperature(plane, climate, pos);
     s32 petRatio = cell->humidityPreference;
     // TODO: translate biotemp and petRatio into range of name array
-    s32 biotempRemap = sqrt(biotemp); // [0, 2400] -> [0, 48]
+    s32 biotempRemap = sqrt((double)biotemp); // [0, 2400] -> [0, 48]
     biotempRemap = remap_int(biotempRemap, 2, 48, 1, 5);
     biotemp = biotemp < -1750 ? 0 : biotempRemap;
     if (cell->height < 0) {
@@ -386,11 +386,3 @@ const char *plane_getCellLifezoneName(const Plane *plane, const Climate *climate
     CLAMP(biotemp, 0, 5);
     return lifezoneNames[petRatio][biotemp];
 }
-
-
-
-
-
-
-
-
